+++ core/src/test/java/uk/gov/justice/services/core/it/AllEventsHandlerIT.java
@@ -19,16 +19,16 @@ import uk.gov.justice.services.core.annotation.Adapter;
 import uk.gov.justice.services.core.annotation.Handles;
 import uk.gov.justice.services.core.annotation.ServiceComponent;
 import uk.gov.justice.services.core.cdi.LoggerProducer;
-import uk.gov.justice.services.core.dispatcher.AsynchronousDispatcher;
-import uk.gov.justice.services.core.dispatcher.AsynchronousDispatcherProducer;
 import uk.gov.justice.services.core.dispatcher.DispatcherCache;
 import uk.gov.justice.services.core.dispatcher.DispatcherFactory;
 import uk.gov.justice.services.core.dispatcher.RequesterProducer;
 import uk.gov.justice.services.core.dispatcher.ServiceComponentObserver;
-import uk.gov.justice.services.core.dispatcher.SynchronousDispatcherProducer;
 import uk.gov.justice.services.core.enveloper.Enveloper;
 import uk.gov.justice.services.core.extension.AnnotationScanner;
 import uk.gov.justice.services.core.extension.BeanInstantiater;
+import uk.gov.justice.services.core.interceptor.InterceptorCache;
+import uk.gov.justice.services.core.interceptor.InterceptorChainProcessor;
+import uk.gov.justice.services.core.interceptor.InterceptorChainProcessorProducer;
 import uk.gov.justice.services.core.it.util.repository.StreamBufferOpenEjbAwareJdbcRepository;
 import uk.gov.justice.services.core.it.util.repository.StreamStatusOpenEjbAwareJdbcRepository;
 import uk.gov.justice.services.core.jms.JmsDestinations;
@@ -63,7 +63,7 @@ public class AllEventsHandlerIT {
     private static final String EVENT_ABC = "test.event-abc";
 
     @Inject
-    private AsynchronousDispatcher asyncDispatcher;
+    private InterceptorChainProcessor interceptorChainProcessor;
 
     @Inject
     private AbcEventHandler abcEventHandler;
@@ -76,10 +76,12 @@ public class AllEventsHandlerIT {
             AbcEventHandler.class,
             AllEventsHandler.class,
             AnnotationScanner.class,
-            AsynchronousDispatcherProducer.class,
-            SynchronousDispatcherProducer.class,
             RequesterProducer.class,
             ServiceComponentObserver.class,
+
+            InterceptorChainProcessorProducer.class,
+            InterceptorChainProcessor.class,
+            InterceptorCache.class,
 
             SenderProducer.class,
             JmsSenderFactory.class,
@@ -117,7 +119,7 @@ public class AllEventsHandlerIT {
     public void shouldHandleEventByName() {
 
         UUID metadataId = randomUUID();
-        asyncDispatcher.dispatch(envelope()
+        interceptorChainProcessor.process(envelope()
                 .with(metadataOf(metadataId, EVENT_ABC)
                         .withStreamId(randomUUID())
                         .withVersion(1l)).build());
@@ -130,7 +132,7 @@ public class AllEventsHandlerIT {
     public void shouldHandleEventByTheAllEventsHandlerIfNamedHandlerNotFound() {
 
         UUID metadataId = randomUUID();
-        asyncDispatcher.dispatch(envelope()
+        interceptorChainProcessor.process(envelope()
                 .with(metadataOf(metadataId, "some.unregistered.event")
                         .withStreamId(randomUUID())
                         .withVersion(1l)).build());