+++ generators/messaging-adapter-generator/src/test/java/uk/gov/justice/raml/jms/core/JmsEndpointGeneratorTest.java
@@ -41,7 +41,7 @@ import uk.gov.justice.services.adapter.messaging.JmsProcessor;
 import uk.gov.justice.services.adapter.messaging.JsonSchemaValidationInterceptor;
 import uk.gov.justice.services.core.annotation.Adapter;
 import uk.gov.justice.services.core.annotation.Component;
-import uk.gov.justice.services.core.dispatcher.AsynchronousDispatcher;
+import uk.gov.justice.services.core.interceptor.InterceptorChainProcessor;
 import uk.gov.justice.services.generators.test.utils.BaseGeneratorTest;
 import uk.gov.justice.services.messaging.JsonEnvelope;
 
@@ -79,12 +79,13 @@ import org.slf4j.Logger;
 public class JmsEndpointGeneratorTest extends BaseGeneratorTest {
     private static final String BASE_PACKAGE = "uk.test";
     private static final String BASE_PACKAGE_FOLDER = "/uk/test";
+    private static final String INTERCEPTOR_CHAIN_PROCESSOR = "interceptorChainProcessor";
 
     @Mock
     JmsProcessor jmsProcessor;
 
     @Mock
-    AsynchronousDispatcher dispatcher;
+    InterceptorChainProcessor interceptorChainProcessor;
 
     @Before
     public void setup() throws Exception {
@@ -364,15 +365,15 @@ public class JmsEndpointGeneratorTest extends BaseGeneratorTest {
     }
 
     @Test
-    public void shouldCreateJmsEndpointWithAnnotatedDispatcherProperty() throws Exception {
+    public void shouldCreateJmsEndpointWithAnnotatedInterceptorChainProcessorProperty() throws Exception {
         generator.run(raml().withDefaultMessagingResource().build(), configurationWithBasePackage(BASE_PACKAGE, outputFolder, emptyMap()));
 
         Class<?> clazz = getJmsListenerClass(BASE_PACKAGE, "SomecontextControllerCommandJmsListener");
-        Field dispatcherField = clazz.getDeclaredField("dispatcher");
-        assertThat(dispatcherField, not(nullValue()));
-        assertThat(dispatcherField.getType(), CoreMatchers.equalTo((AsynchronousDispatcher.class)));
-        assertThat(dispatcherField.getAnnotations(), arrayWithSize(1));
-        assertThat(dispatcherField.getAnnotation(Inject.class), not(nullValue()));
+        Field chainProcessField = clazz.getDeclaredField(INTERCEPTOR_CHAIN_PROCESSOR);
+        assertThat(chainProcessField, not(nullValue()));
+        assertThat(chainProcessField.getType(), CoreMatchers.equalTo((InterceptorChainProcessor.class)));
+        assertThat(chainProcessField.getAnnotations(), arrayWithSize(1));
+        assertThat(chainProcessField.getAnnotation(Inject.class), not(nullValue()));
     }
 
     @Test
@@ -678,7 +679,7 @@ public class JmsEndpointGeneratorTest extends BaseGeneratorTest {
         JsonEnvelope envelope = envelope().build();
         consumerCaptor.getValue().accept(envelope);
 
-        verify(dispatcher).dispatch(envelope);
+        verify(interceptorChainProcessor).process(envelope);
     }
 
     @Test
@@ -729,7 +730,7 @@ public class JmsEndpointGeneratorTest extends BaseGeneratorTest {
     private Object instantiate(Class<?> resourceClass) throws InstantiationException, IllegalAccessException {
         Object resourceObject = resourceClass.newInstance();
         setField(resourceObject, "jmsProcessor", jmsProcessor);
-        setField(resourceObject, "dispatcher", dispatcher);
+        setField(resourceObject, INTERCEPTOR_CHAIN_PROCESSOR, interceptorChainProcessor);
         return resourceObject;
     }