+++ generators/rest-adapter-generator/src/test/java/uk/gov/justice/api/resource/DefaultUsersUserIdResourceIT.java
@@ -26,8 +26,7 @@ import uk.gov.justice.services.common.converter.jackson.ObjectMapperProducer;
 import uk.gov.justice.services.core.cdi.LoggerProducer;
 import uk.gov.justice.services.core.json.JsonSchemaLoader;
 import uk.gov.justice.services.core.json.JsonSchemaValidator;
-import uk.gov.justice.services.generators.test.utils.dispatcher.AsynchronousRecordingDispatcher;
-import uk.gov.justice.services.generators.test.utils.dispatcher.SynchronousRecordingDispatcher;
+import uk.gov.justice.services.generators.test.utils.interceptor.RecordingInterceptorChainProcessor;
 import uk.gov.justice.services.messaging.JsonEnvelope;
 import uk.gov.justice.services.messaging.JsonObjectEnvelopeConverter;
 
@@ -68,10 +67,7 @@ public class DefaultUsersUserIdResourceIT {
     private static String BASE_URI;
 
     @Inject
-    AsynchronousRecordingDispatcher asyncDispatcher;
-
-    @Inject
-    SynchronousRecordingDispatcher syncDispatcher;
+    RecordingInterceptorChainProcessor interceptorChainProcessor;
 
     @Inject
     CommonProviders commonProviders;
@@ -96,8 +92,7 @@ public class DefaultUsersUserIdResourceIT {
             RestProcessor.class,
             RestProcessorProducer.class,
             RestEnvelopeBuilderFactory.class,
-            AsynchronousRecordingDispatcher.class,
-            SynchronousRecordingDispatcher.class,
+            RecordingInterceptorChainProcessor.class,
             ObjectMapperProducer.class,
             JsonObjectEnvelopeConverter.class,
             CommonProviders.class,
@@ -132,7 +127,7 @@ public class DefaultUsersUserIdResourceIT {
                 .path("/users/567-8910")
                 .post(entity("{\"userUrn\" : \"1234\"}", CREATE_USER_MEDIA_TYPE));
 
-        JsonEnvelope jsonEnvelope = asyncDispatcher.awaitForEnvelopeWithPayloadOf("userId", "567-8910");
+        JsonEnvelope jsonEnvelope = interceptorChainProcessor.awaitForEnvelopeWithPayloadOf("userId", "567-8910");
         assertThat(jsonEnvelope.metadata().name(), is("people.create-user"));
         assertThat(jsonEnvelope.payloadAsJsonObject().getString("userId"), is("567-8910"));
         assertThat(jsonEnvelope.payloadAsJsonObject().getString("userUrn"), is("1234"));
@@ -158,7 +153,7 @@ public class DefaultUsersUserIdResourceIT {
 
     @Test
     public void shouldReturn200ResponseContainingUserDataReturnedByDispatcher() {
-        syncDispatcher.setupResponse("userId", "4444-5556",
+        interceptorChainProcessor.setupResponse("userId", "4444-5556",
                 envelope().with(metadataWithDefaults()).withPayloadOf("user1234", "userName").build());
 
 
@@ -179,7 +174,7 @@ public class DefaultUsersUserIdResourceIT {
                 .path("/users/4444-9876")
                 .post(entity("{\"userUrn\" : \"5678\"}", UPDATE_USER_MEDIA_TYPE));
 
-        JsonEnvelope envelope = asyncDispatcher.awaitForEnvelopeWithPayloadOf("userId", "4444-9876");
+        JsonEnvelope envelope = interceptorChainProcessor.awaitForEnvelopeWithPayloadOf("userId", "4444-9876");
         assertThat(envelope.metadata().name(), is("people.update-user"));
         assertThat(envelope.payloadAsJsonObject().getString("userId"), is("4444-9876"));
         assertThat(envelope.payloadAsJsonObject().getString("userUrn"), is("5678"));
@@ -191,7 +186,7 @@ public class DefaultUsersUserIdResourceIT {
                 .path("/users/4444-9877")
                 .post(entity("", LINK_USER_MEDIA_TYPE));
 
-        JsonEnvelope envelope = asyncDispatcher.awaitForEnvelopeWithPayloadOf("userId", "4444-9877");
+        JsonEnvelope envelope = interceptorChainProcessor.awaitForEnvelopeWithPayloadOf("userId", "4444-9877");
         assertThat(envelope.metadata().name(), is("people.link-user"));
         assertThat(envelope.payloadAsJsonObject().getString("userId"), is("4444-9877"));
     }
@@ -202,14 +197,14 @@ public class DefaultUsersUserIdResourceIT {
                 .path("/users/4444-5555")
                 .header("Accept", "application/vnd.people.user+json")
                 .get();
-        JsonEnvelope envelope = syncDispatcher.awaitForEnvelopeWithPayloadOf("userId", "4444-5555");
+        JsonEnvelope envelope = interceptorChainProcessor.awaitForEnvelopeWithPayloadOf("userId", "4444-5555");
         assertThat(envelope.payloadAsJsonObject().getString("userId"), is("4444-5555"));
         assertThat(envelope.metadata().name(), is("people.get-user"));
     }
 
     @Test
     public void shouldDispatchGetUserCommandWithOtherMediaType() throws Exception {
-        syncDispatcher.setupResponse("userId", "4444-5555", envelope().with(metadataWithDefaults()).build());
+        interceptorChainProcessor.setupResponse("userId", "4444-5555", envelope().with(metadataWithDefaults()).build());
 
         Response response = create(BASE_URI)
                 .path("/users/4444-5555")
@@ -217,7 +212,7 @@ public class DefaultUsersUserIdResourceIT {
                 .get();
 
         assertThat(response.getStatus(), is(OK.getStatusCode()));
-        JsonEnvelope envelope = syncDispatcher.awaitForEnvelopeWithPayloadOf("userId", "4444-5555");
+        JsonEnvelope envelope = interceptorChainProcessor.awaitForEnvelopeWithPayloadOf("userId", "4444-5555");
         assertThat(envelope.payloadAsJsonObject().getString("userId"), is("4444-5555"));
         assertThat(envelope.metadata().name(), is("people.get-user-summary"));
     }
@@ -236,7 +231,7 @@ public class DefaultUsersUserIdResourceIT {
 
     @Test
     public void shouldReturnResponseWithContentType() {
-        syncDispatcher.setupResponse("userId", "4444-5556", envelope().with(metadataWithDefaults()).build());
+        interceptorChainProcessor.setupResponse("userId", "4444-5556", envelope().with(metadataWithDefaults()).build());
 
         Response response = create(BASE_URI)
                 .path("/users/4444-5556")
@@ -248,7 +243,7 @@ public class DefaultUsersUserIdResourceIT {
 
     @Test
     public void shouldReturnResponseWithSecondContentType() {
-        syncDispatcher.setupResponse("userId", "4444-5556", envelope().with(metadataWithDefaults()).build());
+        interceptorChainProcessor.setupResponse("userId", "4444-5556", envelope().with(metadataWithDefaults()).build());
 
         Response response = create(BASE_URI)
                 .path("/users/4444-5556")
@@ -260,7 +255,7 @@ public class DefaultUsersUserIdResourceIT {
 
     @Test
     public void shouldDispatchUsersQueryWithQueryParams() throws Exception {
-        syncDispatcher.setupResponse("lastname", "Smith", envelope().with(metadataWithDefaults()).build());
+        interceptorChainProcessor.setupResponse("lastname", "Smith", envelope().with(metadataWithDefaults()).build());
 
         Response response = create(BASE_URI)
                 .path("/users")
@@ -273,7 +268,7 @@ public class DefaultUsersUserIdResourceIT {
                 .get();
 
         assertThat(response.getStatus(), is(OK.getStatusCode()));
-        JsonEnvelope jsonEnvelope = syncDispatcher.awaitForEnvelopeWithPayloadOf("lastname", "Smith");
+        JsonEnvelope jsonEnvelope = interceptorChainProcessor.awaitForEnvelopeWithPayloadOf("lastname", "Smith");
         assertThat(jsonEnvelope.metadata().name(), is("people.search-users"));
 
         final JsonObject payload = (JsonObject) jsonEnvelope.payload();
@@ -298,7 +293,7 @@ public class DefaultUsersUserIdResourceIT {
 
     @Test
     public void shouldReturn200WhenOptionalParameterIsNotProvided() throws Exception {
-        syncDispatcher.setupResponse("lastname", "lastname", envelope().with(metadataWithDefaults()).build());
+        interceptorChainProcessor.setupResponse("lastname", "lastname", envelope().with(metadataWithDefaults()).build());
 
         Response response = create(BASE_URI)
                 .path("/users")